Changed implementation of DataParser's class method DeserializeJson in order to deal with missing fields within json documents. It required also to make int, long and double properties of the Request's class nullable to exclude them from deserialization if the fileds are missing (JSON.NET passes a default value to the property if the field in the json file is not present).

+++ BootcampCoreServices/ViewModel/ReportsGenerator.cs
@@ -22,13 +22,13 @@ namespace BootcampCoreServices.ViewModel
         // c
         public static double TotalValueOfRequests(IEnumerable<Request> requests)
         {
-            return requests.Sum(t => t.Price * t.Quantity);
+            return (double) requests.Sum(t => t.Price * t.Quantity);
         }
 
         // d
         public static double TotalValueOfRequests(IEnumerable<Request> requests, string clientId)
         {
-            return requests.Where(c => c.ClientId == clientId).Sum(t => t.Price * t.Quantity);
+            return (double) requests.Where(c => c.ClientId == clientId).Sum(t => t.Price * t.Quantity);
         }
 
         // e
